At long last; moved single_step out of debug/debug_utils.
It now resides in name_core/simulator_helpers.
