+++ name-core/src/debug/exception_handler.rs
@@ -3,10 +3,9 @@ use crate::{
     structs::{OperatingSystem, ProgramState},
 };
 
-use crate::debug::simulator_helpers::generate_err;
+use crate::simulator_helpers::generate_err;
 
 use super::debug_utils::DebuggerState;
-//use name_core::debug::
 
 /// The exception handler is invoked whenever an exception has occurred.
 /// Some common exceptions include breakpoints, syscalls, and arithmetic overflow.
