+++ name-core/src/lib.rs
@@ -6,6 +6,7 @@ pub mod exception;
 pub mod implementations;
 pub mod instruction;
 pub mod parse;
+pub mod simulator_helpers;
 pub mod structs;
 pub mod syscalls;
 pub mod traits;
